# web/src/features/docs/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexus Reach Seedance 2.0 API",
    "version": "2026-07-24.1",
    "description": "Nexus Reach 对外提供 Seedance 2.0 国内视频生成、素材资产以及移动视频素材库/真人认证接口。本 OpenAPI 文件共列出 15 个 paths：原有 9 个 Seedance 国内接口保持不变，另新增 6 个移动视频接口。所有公开请求统一通过 Nexus Reach API Gateway 访问，并使用 Nexus Reach API Key 鉴权；不要将上游供应商域名作为生产请求地址。移动视频接口由独立渠道转发，供应商响应的 HTTP 状态和 JSON 内容由网关透传；成功响应包含 `requestId`、`state`、`errorCode`、`errorMessage` 和 `body`，调用方应确认 `state` 为 `OK` 后再读取 `body`。"
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/v1/video/generations": {
      "post": {
        "tags": ["视频生成"],
        "summary": "创建 Seedance 2.0 视频任务",
        "operationId": "createSeedanceDomesticVideo",
        "description": "这是本 Seedance 国内客户契约的唯一创建入口，请使用 `Content-Type: application/json`，不要改用 OpenAI/Sora 的 `POST /v1/videos` multipart 协议。`prompt` 为必填提示词；图片、视频和音频输入放在 `metadata.content` 中，音频不能作为唯一输入。使用移动视频素材库资源时，应先调用单素材查询接口，确认 `body.status` 为 `ACTIVE`，再将最新的 `body.assetUrl` 写入对应媒体对象的 `url` 字段；不要把移动视频素材 ID 直接拼成 `asset://<assetId>`。提示词应按媒体在 `metadata.content` 中出现的顺序使用“图片1、图片2、视频1”等称呼。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoGenerationRequest"
              },
              "example": {
                "model": "doubao-seedance-2-0-260128",
                "prompt": "图片1中的虚拟人物在海边自然行走，镜头缓慢向前推进。",
                "duration": 4,
                "resolution": "720p",
                "metadata": {
                  "content": [
                    {
                      "type": "image_url",
                      "image_url": {
                        "url": "https://provider.example.invalid/assets/reference.png?signature=example"
                      },
                      "role": "reference_image"
                    }
                  ],
                  "ratio": "adaptive",
                  "generate_audio": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "任务创建成功。请保存返回的公开任务 ID，用于后续查询。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoSubmitResponse"
                },
                "example": {
                  "id": "task_01jzseedanceexample",
                  "task_id": "task_01jzseedanceexample",
                  "object": "video",
                  "model": "doubao-seedance-2-0-260128",
                  "status": "queued",
                  "progress": 0,
                  "created_at": 1784169600
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或内容未通过校验。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                },
                "example": {
                  "code": "invalid_request",
                  "message": "prompt is required",
                  "data": null
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/VideoUnauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "413": {
            "description": "请求体超过网关允许的大小。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskErrorResponse"
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/TaskInternalError" },
          "502": { "$ref": "#/components/responses/TaskUpstreamError" },
          "503": {
            "description": "当前账号分组没有可用的国内 Seedance 渠道。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Nexus Reach API Key，例如：`Authorization: Bearer sk-your-key`。"
      }
    },
    "responses": {
...
      "VideoUnauthorized": {
        "description": "视频接口未提供 Token，或 Token 无效。调用方应以 HTTP 401 判定鉴权失败；`request id` 仅用于问题排查。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "message": "Invalid token (request id: example-request-id)",
                "type": "new_api_error",
                "code": ""
              }
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "当前 API Key 无权访问该接口或资源。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "message": "permission denied",
                "type": "permission_error",
                "code": "permission_denied"
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "请求频率超过限制，请稍后重试。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "error": {
                "message": "rate limit exceeded",
                "type": "rate_limit_error",
                "code": "rate_limit_exceeded"
              }
            }
          }
        }
      },
      "TaskInternalError": {
        "description": "视频任务服务内部错误。视频创建和查询控制器返回裸 TaskError 结构。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/TaskErrorResponse" },
            "example": {
              "code": "internal_error",
              "message": "internal server error",
              "data": null
            }
          }
        }
      },
      "TaskUpstreamError": {
        "description": "视频生成服务返回了无法解析或不完整的响应。视频控制器返回裸 TaskError 结构。",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/TaskErrorResponse" },
            "example": {
              "code": "invalid_response",
              "message": "invalid upstream task id",
              "data": null
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ContentItem": {
        "type": "object",
        "description": "`metadata.content` 中的媒体输入项。`type` 与对应媒体字段必须匹配；媒体编号按同类型项目在数组中的出现顺序计算。",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["image_url", "video_url", "audio_url"],
            "description": "内容类型。"
          },
          "image_url": {
            "$ref": "#/components/schemas/ContentMedia"
          },
          "video_url": {
            "$ref": "#/components/schemas/ContentMedia"
          },
          "audio_url": {
            "$ref": "#/components/schemas/ContentMedia"
          },
          "role": {
            "type": "string",
            "description": "可选的供应商媒体语义角色。常见值包括 reference_image、first_frame、reference_video 和 reference_audio；网关会透传该字段。"
          }
        },
        "oneOf": [
          {
            "required": ["type", "image_url"],
            "properties": { "type": { "enum": ["image_url"] } }
          },
          {
            "required": ["type", "video_url"],
            "properties": { "type": { "enum": ["video_url"] } }
          },
          {
            "required": ["type", "audio_url"],
            "properties": { "type": { "enum": ["audio_url"] } }
          }
        ]
      },
      "VideoGenerationMetadata": {
        "type": "object",
        "description": "转发给 Seedance 上游的扩展参数。媒体输入应放在 `content` 中。",
        "properties": {
          "content": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/ContentItem"
            },
            "description": "可选的图片、视频或音频输入。音频不能作为唯一输入。"
          },
          "ratio": {
            "type": "string",
            "enum": ["adaptive", "16:9", "4:3", "1:1", "3:4", "9:16", "21:9"],
            "default": "adaptive",
            "description": "输出宽高比。"
          },
          "generate_audio": {
            "type": "boolean",
            "default": true,
            "description": "是否生成音频。"
          }
        },
        "additionalProperties": true
      },
      "VideoGenerationRequest": {
        "type": "object",
        "description": "当前国内 Seedance 通道使用顶层 `prompt`，并通过 `metadata.content` 提供可选媒体输入。",
        "properties": {
          "model": {
            "type": "string",
            "enum": ["doubao-seedance-2-0-260128"],
            "default": "doubao-seedance-2-0-260128",
            "description": "模型名称。"
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\S",
            "description": "提示词。引用媒体时使用“图片1、视频1”等顺序称呼，不要填写素材 ID。"
          },
          "image": {
            "type": "string",
            "minLength": 1,
            "description": "单张图片 URL 简写。需要指定 `role` 时请改用 `metadata.content`。"
          },
          "images": {
            "type": "array",
            "minItems": 1,
            "maxItems": 9,
            "items": {
              "type": "string",
              "minLength": 1,
              "pattern": "\\S"
            },
            "description": "多张图片 URL 简写，最多 9 张。需要指定 `role` 时请改用 `metadata.content`。"
          },
          "resolution": {
            "type": "string",
            "enum": ["720p", "1080p", "4K"],
            "default": "720p",
            "description": "输出分辨率。可选 720p、1080p 或 4K。"
          },
          "duration": {
            "type": "integer",
            "minimum": 4,
            "maximum": 15,
            "default": 5,
            "description": "输出时长，单位为秒，支持 4～15 的整数。"
          },
          "metadata": {
            "$ref": "#/components/schemas/VideoGenerationMetadata"
          }
        },
        "required": ["model", "prompt"]
      },
      "VideoSubmitResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "公开任务 ID。"
          },
          "task_id": {
            "type": "string",
            "description": "兼容任务 ID 字段，与 id 相同。"
          },
          "object": {
            "type": "string",
            "enum": ["video"]
          },
          "model": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["queued", "in_progress", "completed", "failed"]
          },
          "progress": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "created_at": {
            "type": "integer",
            "description": "Unix 时间戳。"
          }
        },
        "required": [
          "id",
          "task_id",
          "object",
          "model",
          "status",
          "progress",
          "created_at"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "code": {
                "type": "string"
              }
            },
            "required": ["message", "code"]
          }
        },
        "required": ["error"]
      },
      "TaskErrorResponse": {
        "type": "object",
        "description": "视频任务控制器返回的错误格式。它与鉴权中间件返回的 error 包装格式不同。",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": ["code", "message", "data"],
        "additionalProperties": false
      }
    }
  }
}